+++ robot_hat/data_types/config/__init__.py
@@ -1,0 +1,3 @@
+from robot_hat.data_types.config.lidar import RPLidarC1Config
+
+__all__ = ["RPLidarC1Config"]