+++ robot_hat/sensors/__init__.py
@@ -1,0 +1,3 @@
+from robot_hat.sensors.lidar import RPLidarC1
+
+__all__ = ["RPLidarC1"]